refactor: move auth feature, remove unnecessary class
