--- src/main/java/com/lpvs/auth/MyAuthenticationSuccessHandler.java
@@ -1,41 +1,0 @@
-/**
- * Copyright 2023 Basaeng, kyudori, hwan5180, quswjdgma83
- *
- * Use of this source code is governed by a MIT license that can be
- * found in the LICENSE file.
- */
-
-package com.lpvs.auth;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.util.UriComponentsBuilder;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-
-@Component
-public class MyAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
-    @Value("${frontend.main-page.url:}")
-    private String frontendMainPageUrl;
-
-    private String REDIRECT_URI = frontendMainPageUrl+"/login/callback";
-
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
-        OAuth2User oAuth2User = (OAuth2User) authentication.getPrincipal();
-        System.out.println("oAuth2User = " + oAuth2User);
-
-        response.sendRedirect(UriComponentsBuilder.fromUriString(REDIRECT_URI)
-                .queryParam("accessToken", "accessToken")
-                .queryParam("refreshToken", "refreshToken")
-                .build().encode(StandardCharsets.UTF_8).toUriString());
-    }
-
-}
