+++ src/main/java/com/lpvs/entity/auth/SecurityConfig.java
@@ -5,18 +5,30 @@
  * found in the LICENSE file.
  */
 
-package com.lpvs.auth;
+package com.lpvs.entity.auth;
 
 import lombok.RequiredArgsConstructor;
+
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.util.UriComponentsBuilder;
 
 @EnableWebSecurity
 @RequiredArgsConstructor
@@ -48,7 +60,23 @@ public class SecurityConfig {
                 .anyRequest().permitAll()
                 .and()
                 .oauth2Login()
-                .successHandler(new MyAuthenticationSuccessHandler())
+                .successHandler(new AuthenticationSuccessHandler() {
+                    @Value("${frontend.main-page.url:}")
+                    private String frontendMainPageUrl;
+
+                    private String REDIRECT_URI = frontendMainPageUrl+"/login/callback";
+
+                    @Override
+                    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
+                        OAuth2User oAuth2User = (OAuth2User) authentication.getPrincipal();
+                        System.out.println("oAuth2User = " + oAuth2User);
+
+                        response.sendRedirect(UriComponentsBuilder.fromUriString(REDIRECT_URI)
+                                .queryParam("accessToken", "accessToken")
+                                .queryParam("refreshToken", "refreshToken")
+                                .build().encode(StandardCharsets.UTF_8).toUriString());
+                    }
+                })
                 .defaultSuccessUrl(frontendMainPageUrl, true)
                 .userInfoEndpoint()
                 .userService(oAuthService);
